+++ cognite/client/_api/units.py
@@ -6,12 +6,11 @@ from functools import cache
 from itertools import chain
 from typing import TYPE_CHECKING, Literal, overload
 
+from cognite.client._api.unit_system import UnitSystemAPI
 from cognite.client._api_client import APIClient
 from cognite.client.data_classes.units import (
     Unit,
     UnitList,
-    UnitSystem,
-    UnitSystemList,
 )
 from cognite.client.utils._auxiliary import remove_duplicates_keep_order
 from cognite.client.utils._identifier import IdentifierSequence
@@ -220,24 +219,3 @@ class UnitAPI(APIClient):
                 >>> res = client.units.list()
         """
         return self._list(method="GET", list_cls=UnitList, resource_cls=Unit)
-
-
-class UnitSystemAPI(APIClient):
-    _RESOURCE_PATH = "/units/systems"
-
-    def list(self) -> UnitSystemList:
-        """`List all supported unit systems <https://developer.cognite.com/api#tag/Unit-Systems/operation/listUnitSystems>`_
-
-        Returns:
-            UnitSystemList: List of unit systems
-
-        Examples:
-
-            List all supported unit systems in CDF:
-
-                >>> from cognite.client import CogniteClient
-                >>> client = CogniteClient()
-                >>> res = client.units.systems.list()
-
-        """
-        return self._list(method="GET", list_cls=UnitSystemList, resource_cls=UnitSystem)
